Extract abstract superClass entityEditorTool
